requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
torch==2.5.1
torchvision==0.20.1
pillow==11.0.0
numpy==1.26.4
tqdm==4.66.1
python-multipart==0.0.20
git+https://github.com/openai/CLIP.git
python-dotenv==1.2.1
sentence-transformers>=3.0.0